edtech-ai/backend/requirements.txt
fastapi[standard]
uvicorn
sqlalchemy
psycopg2-binary
alembic
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
pytest
httpx
